Config update

+++ src/configuration/src/lib.rs
@@ -19,7 +19,7 @@ pub enum TransitionType {
 #[derive(Serialize, Deserialize)]
 pub struct DistributionSpecial {
 	pub uuid: Vec<String>,
-	pub density: f32, // TODO: change to f64
+	pub density: f64,
 	#[serde(rename(deserialize = "_comment"))]
 	pub comment: String,
 }
@@ -27,7 +27,7 @@ pub struct DistributionSpecial {
 #[derive(Serialize, Deserialize)]
 pub struct TransitionSpecial {
 	pub uuid: Vec<String>,
-	pub width: f32,
+	pub width: f64,
 	#[serde(rename(deserialize = "_comment"))]
 	pub comment: String,
 }
@@ -36,7 +36,7 @@ pub struct TransitionSpecial {
 pub struct Distribution {
 	#[serde(rename(deserialize = "type"))]
 	pub distribution_type: DistributionType,
-	pub density: f32, // TODO: change to f64
+	pub density: f64,
 	pub special: Vec<DistributionSpecial>,
 }
 
@@ -45,21 +45,21 @@ pub struct Transition {
 	#[serde(rename(deserialize = "type"))]
 	pub transitions_type: TransitionType,
 	#[serde(rename(deserialize = "doorwayin"))]
-	pub doorway_in: f32,
+	pub doorway_in: f64,
 	#[serde(rename(deserialize = "doorwayout"))]
-	pub doorway_out: f32,
+	pub doorway_out: f64,
 	pub special: Vec<TransitionSpecial>,
 }
 
 #[derive(Serialize, Deserialize)]
 pub struct Modeling {
-	pub step: f32,
+	pub step: f64,
 	#[serde(rename(deserialize = "speed_max"))]
-	pub max_speed: f32,
+	pub max_speed: f64,
 	#[serde(rename(deserialize = "density_max"))]
-	pub max_density: f32,
+	pub max_density: f64,
 	#[serde(rename(deserialize = "density_min"))]
-	pub min_density: f32,
+	pub min_density: f64,
 }
 
 #[derive(Serialize, Deserialize)]
